Implement logging to SD card
Measured speed data is saved to CSV files in the CDM324 directory,
with the logging rate matching the display update rate. Upon system
startup, a new file is created, and its name is incremented
numerically. Following idle periods, the data is flushed to the
storage card to prevent data loss.